src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/hover-active-drag-distributed-nodes-expected.txt

CONSOLE ERROR: line 56: Uncaught TypeError: host.createShadowRoot is not a function
Top-level text Nested text
Other div
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
